+++ src/dbally/views/exceptions.py
@@ -1,0 +1,26 @@
+from typing import Optional
+
+from dbally.exceptions import DbAllyError
+
+
+class IQLGenerationError(DbAllyError):
+    """
+    Exception for when an error occurs while generating IQL for a view.
+    """
+
+    def __init__(
+        self,
+        view_name: str,
+        filters: Optional[str] = None,
+        aggregation: Optional[str] = None,
+    ) -> None:
+        """
+        Args:
+            view_name: Name of the view that caused the error.
+            filters: Filters generated by the view.
+            aggregation: Aggregation generated by the view.
+        """
+        super().__init__(f"Error while generating IQL for view {view_name}")
+        self.view_name = view_name
+        self.filters = filters
+        self.aggregation = aggregation
