+++ src/dbally/views/structured.py
@@ -4,41 +4,17 @@ from typing import Dict, List, Optional
 
 from dbally.audit.event_tracker import EventTracker
 from dbally.collection.results import ViewExecutionResult
-from dbally.exceptions import DbAllyError
 from dbally.iql import IQLQuery
 from dbally.iql._exceptions import IQLError
 from dbally.iql_generator.iql_generator import IQLGenerator
 from dbally.iql_generator.prompt import UnsupportedQueryError
 from dbally.llms.base import LLM
 from dbally.llms.clients.base import LLMOptions
+from dbally.views.exceptions import IQLGenerationError
 from dbally.views.exposed_functions import ExposedFunction
 
 from ..similarity import AbstractSimilarityIndex
 from .base import BaseView, IndexLocation
-
-
-# TODO: Move this error to generators
-class IQLGenerationError(DbAllyError):
-    """
-    Exception for when an error occurs while executing a view.
-    """
-
-    def __init__(
-        self,
-        view_name: str,
-        filters: Optional[str] = None,
-        aggregation: Optional[str] = None,
-    ) -> None:
-        """
-        Args:
-            view_name: Name of the view that caused the error.
-            filters: Filters generated by the view.
-            aggregation: Aggregation generated by the view.
-        """
-        super().__init__(f"Error executing view '{view_name}'.")
-        self.view_name = view_name
-        self.filters = filters
-        self.aggregation = aggregation
 
 
 class BaseStructuredView(BaseView):
